Moved catsinstances to cats.instances, this will be more consistent: deprecated old version
