+++ parsley-cats/shared/src/main/scala-2.13+/parsley/cats/MonoidKForParsley.scala
@@ -1,7 +1,10 @@
 /* SPDX-FileCopyrightText: © 2022 Parsley Cats Contributors <https://github.com/j-mie6/parsley-cats/graphs/contributors>
  * SPDX-License-Identifier: BSD-3-Clause
  */
-package parsley
+package parsley.cats
+
+import parsley.Parsley
+import parsley.combinator
 
 import cats.{Functor, MonoidK}
 
@@ -12,6 +15,6 @@ private [parsley] trait MonoidKForParsley extends MonoidK[Parsley] {
 
     // MonoidK Overrides
     override def sum[A, B](mx: Parsley[A], my: Parsley[B])(implicit F: Functor[Parsley]): Parsley[Either[A,B]] = mx <+> my
-    override def combineAllK[A](ps: TraversableOnce[Parsley[A]]): Parsley[A] = combinator.choice(ps.toIterator.toSeq: _*)
-    override def combineAllOptionK[A](ps: TraversableOnce[Parsley[A]]): Option[Parsley[A]] = ps.toIterator.reduceRightOption(_<|>_)
+    override def combineAllK[A](ps: IterableOnce[Parsley[A]]): Parsley[A] = combinator.choice(ps.iterator.toSeq: _*)
+    override def combineAllOptionK[A](ps: IterableOnce[Parsley[A]]): Option[Parsley[A]] = ps.iterator.reduceRightOption(_<|>_)
 }
